# openapi.yaml
openapi: 3.0.3
info:
  title: Mock Interview Buddy Platform API
  description: >
    REST and WebSocket specification expected by the Mock Interview Buddy frontend application.
    Powers real-time collaborative system design whiteboard interviews between interviewers,
    candidates, observers, and administrators.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development backend server
  - url: https://api.mockinterviewbuddy.com
    description: Production backend cluster
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/sessions/{sessionId}/snapshots:
    get:
      tags:
        - Snapshots
      summary: List session snapshots
      description: Lists all captured point-in-time canvas snapshots for the session.
      operationId: listSnapshots
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Saved snapshots
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Snapshot"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      tags:
        - Snapshots
      summary: Save canvas snapshot
      description: >
        Captures the current state of canvas nodes and edges as an immutable snapshot.
        Expected by `PlatformApi.saveSnapshot(sessionId, label)`.
      operationId: saveSnapshot
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - label
              properties:
                label:
                  type: string
                  description: Human-readable name/label for the snapshot
                  example: "Snapshot 1 - Initial Architecture"
      responses:
        "201":
          description: Snapshot saved
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/Snapshot"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v1/sessions/{sessionId}/snapshots/{snapshotId}/restore:
    post:
      tags:
        - Snapshots
      summary: Restore canvas snapshot
      description: >
        Overwrites current session canvas nodes and edges with the saved snapshot state,
        and broadcasts updated state to all connected participants.
        Expected by `PlatformApi.restoreSnapshot(sessionId, snapshotId)`.
      operationId: restoreSnapshot
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: Restored session board
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/SessionBoard"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ============================================================================
  # REAL-TIME WEBSOCKET GATEWAY
  # ============================================================================
  /api/v1/sessions/{sessionId}/ws:
    get:
      tags:
        - Realtime
      summary: Real-time session WebSocket connection
      description: >
        Upgrades HTTP to persistent WebSocket connection. Used by `PlatformApi.subscribe(sessionId, onChange)`.
        Streams live presence (cursors), remote canvas node/edge mutations, chat notifications,
        and board synchronization.
      operationId: sessionRealtimeGateway
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: token
          in: query
          description: Optional auth token when WebSocket client cannot send Authorization headers directly
          required: false
          schema:
            type: string
      responses:
        "101":
          description: WebSocket upgrade handshake successful
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

# ==============================================================================
# COMPONENTS & SCHEMAS
# ==============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter valid JWT token issued on login or session invite join.
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Unique ID of the interview session (e.g. ses_demo)
      schema:
        type: string
        example: "ses_demo"
...
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      description: Unique ID of the snapshot (e.g. snp_001)
      schema:
        type: string
        example: "snp_001"
  schemas:
...
    Role:
      type: string
      enum:
        - interviewer
        - candidate
        - observer
        - admin
      description: User permission tier in the interview platform.
      example: "interviewer"
    SessionStatus:
      type: string
      enum:
        - scheduled
        - active
        - paused
        - ended
      description: Lifecycle state of the interview session.
      example: "active"
    NodeKind:
      type: string
      enum:
        - client
        - gateway
        - service
        - database
        - cache
        - queue
        - storage
        - cdn
        - worker
        - external
      description: Catalog architectural archetype for a canvas node.
      example: "database"
    EdgeProtocol:
      type: string
      enum:
        - http
        - grpc
        - async
        - sql
        - stream
      description: Communication protocol representing connection semantics.
      example: "http"

    # --------------------------------------------------------------------------
    # Domain Models
    # --------------------------------------------------------------------------
...
    Participant:
      type: object
      required:
        - id
        - sessionId
        - userId
        - displayName
        - role
        - color
        - online
      properties:
        id:
          type: string
          description: Participant connection entry ID
          example: "p1"
        sessionId:
          type: string
          description: Associated interview session ID
          example: "ses_demo"
        userId:
          type: string
          description: User ID of participant
          example: "u_interviewer"
        displayName:
          type: string
          description: Display name
          example: "Ada Mensah"
        role:
          $ref: "#/components/schemas/Role"
        color:
          type: string
          description: Assigned participant color hex
          example: "#f0b429"
        online:
          type: boolean
          description: Current live connection status
          example: true
        cursor:
          type: object
          nullable: true
          description: Real-time board cursor coordinates
          properties:
            x:
              type: number
              example: 420
            y:
              type: number
              example: 260
    Session:
      type: object
      required:
        - id
        - title
        - prompt
        - status
        - createdAt
        - createdBy
        - inviteToken
        - durationMinutes
      properties:
        id:
          type: string
          description: Unique session ID
          example: "ses_demo"
        title:
          type: string
          description: Interview session title
          example: "Design a URL shortener"
        prompt:
          type: string
          description: System design problem description and scenario requirements
          example: "Design a highly available URL shortening service handling 10k writes/s and 1M reads/s with analytics."
        status:
          $ref: "#/components/schemas/SessionStatus"
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 creation timestamp
          example: "2026-09-15T12:00:00.000Z"
        createdBy:
          type: string
          description: User ID who created the session
          example: "u_interviewer"
        inviteToken:
          type: string
          description: Shareable invite code for candidates
          example: "inv-7f3a91"
        durationMinutes:
          type: integer
          description: Scheduled session duration in minutes
          example: 60
    CanvasNode:
      type: object
      required:
        - id
        - sessionId
        - kind
        - label
        - notes
        - x
        - y
        - width
        - height
        - version
      properties:
        id:
          type: string
          description: Unique canvas node ID
          example: "nd_gateway1"
        sessionId:
          type: string
          description: Session ID containing this node
          example: "ses_demo"
        kind:
          $ref: "#/components/schemas/NodeKind"
        label:
          type: string
          description: Display name of the architecture component
          example: "API Gateway"
        notes:
          type: string
          description: Notes or configuration details for this component
          example: "Rate limiting and SSL termination"
        x:
          type: number
          description: X coordinate on infinite canvas
          example: 330
        y:
          type: number
          description: Y coordinate on infinite canvas
          example: 120
        width:
          type: number
          description: Rendered width in pixels
          example: 176
        height:
          type: number
          description: Rendered height in pixels
          example: 84
        version:
          type: integer
          description: Monotonically increasing version counter for optimistic concurrency
          example: 1
    CanvasEdge:
      type: object
      required:
        - id
        - sessionId
        - fromNodeId
        - toNodeId
        - label
        - protocol
      properties:
        id:
          type: string
          description: Unique edge ID
          example: "eg_1"
        sessionId:
          type: string
          description: Session ID
          example: "ses_demo"
        fromNodeId:
          type: string
          description: Source canvas node ID
          example: "nd_client"
        toNodeId:
          type: string
          description: Target canvas node ID
          example: "nd_gateway"
        label:
          type: string
          description: Optional protocol or interaction note displayed along the edge
          example: "HTTPS"
        protocol:
          $ref: "#/components/schemas/EdgeProtocol"
    ChatMessage:
      type: object
      required:
        - id
        - sessionId
        - authorId
        - authorName
        - body
        - createdAt
      properties:
        id:
          type: string
          description: Unique chat message ID
          example: "msg_123"
        sessionId:
          type: string
          description: Session ID
          example: "ses_demo"
        authorId:
          type: string
          description: User ID who posted the message
          example: "u_candidate"
        authorName:
          type: string
          description: Author display name
          example: "Noah Chen"
        body:
          type: string
          description: Message text
          example: "Starting with read-heavy assumptions — 100:1 read to write ratio."
        createdAt:
          type: string
          format: date-time
          example: "2026-09-15T12:05:00.000Z"
    AuditEntry:
      type: object
      required:
        - id
        - sessionId
        - actorName
        - action
        - detail
        - createdAt
      properties:
        id:
          type: string
          description: Unique audit log entry ID
          example: "aud_123"
        sessionId:
          type: string
          description: Session ID
          example: "ses_demo"
        actorName:
          type: string
          description: Name of actor who performed the action
          example: "Ada Mensah"
        action:
          type: string
          description: Operation or event identifier (e.g. node.created, session.active)
          example: "node.created"
        detail:
          type: string
          description: Human-readable summary of the action
          example: "gateway · API Gateway"
        createdAt:
          type: string
          format: date-time
          example: "2026-09-15T12:01:00.000Z"
    Snapshot:
      type: object
      required:
        - id
        - sessionId
        - label
        - createdAt
        - nodes
        - edges
      properties:
        id:
          type: string
          description: Unique snapshot ID
          example: "snp_123"
        sessionId:
          type: string
          description: Session ID
          example: "ses_demo"
        label:
          type: string
          description: User-defined or auto-generated snapshot label
          example: "Snapshot 1"
        createdAt:
          type: string
          format: date-time
          example: "2026-09-15T12:20:00.000Z"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/CanvasNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/CanvasEdge"
    SessionBoard:
      type: object
      required:
        - session
        - participants
        - nodes
        - edges
        - messages
        - audit
        - snapshots
      properties:
        session:
          $ref: "#/components/schemas/Session"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/CanvasNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/CanvasEdge"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        audit:
          type: array
          items:
            $ref: "#/components/schemas/AuditEntry"
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/Snapshot"

    # --------------------------------------------------------------------------
    # Request Input Types
    # --------------------------------------------------------------------------
...
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Invalid input parameters"
        details:
          type: object
          nullable: true
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"

  # ============================================================================
  # STANDARD RESPONSES
  # ============================================================================
  responses:
    BadRequestError:
      description: Invalid request payload, missing fields, or validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Authentication is missing, invalid, or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ForbiddenError:
      description: Authenticated user lacks required permissions for this operation/session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundError:
      description: The requested resource (session, node, edge, or snapshot) was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"